% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMatrixMultiplication.R
\name{insertMatrixMultiplication}
\alias{insertMatrixMultiplication}
\title{Insert Matrix multiplication.}
\usage{
insertMatrixMultiplication()
}
\description{
Call this function as an addin to insert \code{ \%\*\% } at the cursor position.
}
