% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPipeOperator.R
\name{insertPipeOperator}
\alias{insertPipeOperator}
\title{Insert Assign and Pipe Operator.}
\usage{
insertPipeOperator()
}
\description{
Call this function as an addin to insert \code{ \%\<\>\% } at the cursor position.
}
