% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeOperator2.R
\name{InsertPipeOperator2}
\alias{InsertPipeOperator2}
\title{Insert \\%\<\>\\%.}
\usage{
InsertPipeOperator2()
}
\description{
Call this function as an addin to insert \code{ \%\<\>\% } at the cursor position.
}
