% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertMatrixMultiplication.R
\name{InsertMatrixMultiplication}
\alias{InsertMatrixMultiplication}
\title{Insert \\%\*\\%.}
\usage{
InsertMatrixMultiplication()
}
\description{
Call this function as an addin to insert \code{ \%\*\% } at the cursor position.
}
